cosmic_utils_enroll::app

Struct AppModel

pub struct AppModel {

    core: Core,
    context_page: ContextPage,
    nav: Model,
    key_binds: HashMap<KeyBind, MenuAction>,
    config: Config,
    status: String,
    device_path: Option<Arc<OwnedObjectPath>>,
    device_proxy: Option<DeviceProxy<'static>>,
    connection: Option<Connection>,
    busy: bool,
    enrolling_finger: Option<Arc<String>>,
    verifying_finger: bool,
    enroll_progress: u32,
    enroll_total_stages: Option<u32>,
    users: Vec<UserOption>,
    selected_user: Option<UserOption>,
    selected_finger: Finger,
    enrolled_fingers: Vec<String>,
    confirm_clear: bool,

}

Application model stores app-specific state

Describes interface and drives its logic

Fields
core: Core

Application state which is managed by the COSMIC runtime.

context_page: ContextPage

Display a context drawer with the designated page if defined.

nav: Model

Contains items assigned to the nav bar panel.

key_binds: HashMap<KeyBind, MenuAction>

Key bindings for the application’s menu bar.

config: Configstatus: Stringdevice_path: Option<Arc<OwnedObjectPath>>device_proxy: Option<DeviceProxy<'static>>connection: Option<Connection>busy: boolenrolling_finger: Option<Arc<String>>verifying_finger: boolenroll_progress: u32enroll_total_stages: Option<u32>users: Vec<UserOption>selected_user: Option<UserOption>selected_finger: Fingerenrolled_fingers: Vec<String>confirm_clear: bool
Implementations

impl AppModel

pub(crate) fn on_cancel_clear(&mut self) -> Task<Action<Message>>

Resets clear state

Returns Task()

pub(crate) fn on_clear_completion( &mut self, res: Result<(), AppError>, ) -> Task<Action<Message>>

After succesfully removal of all prints set status, empties enrolled_fingers

In case of an Error localizes the message and sets status

Returns Task()

pub(crate) fn on_clicked_link(&mut self) -> Task<Action<Message>>

Opens in a browser clicked hyperlink

Returns Task()

pub(crate) fn on_connection_ready( &mut self, conn: Connection, ) -> Task<Action<Message>>

After DBus connection is established searches queries it for fprintd default device

Returns task_find_device(Connection)

pub(crate) fn on_context_page_toggle( &mut self, context_page: ContextPage, ) -> Task<Action<Message>>

Toggles the context page

Returns Task()

pub(crate) fn on_error(&mut self, err: AppError) -> Task<Action<Message>>

Localizes the error and stores it on status resetting everything

Returns Task()

pub(crate) fn on_fingers_listed( &mut self, fingers: Vec<String>, ) -> Task<Action<Message>>

Stores the results of list_fingers_task

Returns Task()

pub(crate) fn on_finger_selected( &mut self, finger: String, ) -> Task<Action<Message>>

If device is not busy compares localized string to fingers and set matching to be the selected one

Returns Task()

pub(crate) fn on_device_found( &mut self, device_info: Option<(OwnedObjectPath, DeviceProxy<'static>)>, ) -> Task<Action<Message>>

Requests users enrolled prints

Returns either Task() or list_fingers_task()

pub(crate) fn on_verify_finger(&mut self) -> Task<Action<Message>>

Called to request verification of the selected print

Returns Task()

pub(crate) fn on_verify_status( &mut self, status: String, done: bool, ) -> Task<Action<Message>>

Handles verification status updates

Returns Task()

pub(crate) fn on_enroll_start( &mut self, total: Option<u32>, ) -> Task<Action<Message>>

Starts the enroll process, set status and enroll options

Returns Task()

pub(crate) fn on_enroll_status( &mut self, status: String, done: bool, ) -> Task<Action<Message>>

Takes responses from Fprintd API and converts them to localized strings

Set status and ends process when it is done

Returns Task()

pub(crate) fn on_enroll_stop(&self) -> Task<Action<Message>>

Sends stop signal to end an ongoing enroll process

Returns either Task() or task_enroll_stop()

pub(crate) fn on_clear_device(&mut self) -> Task<Action<Message>>

Clears all prints for all users

Returns either Task() or task_clear_device()

pub(crate) fn on_delete(&mut self) -> Task<Action<Message>>

Deletes chosen print or users all prints depending on choices from the user

Returns either Task(), task_delete_print() or task_delete_prints()

pub(crate) fn on_delete_complete( &mut self, clear: bool, ) -> Task<Action<Message>>

Set state when deletion of prints was succesful and removes from enrolled_fingers

Returns Task()

pub(crate) fn on_open_link(&mut self, url: String) -> Task<Action<Message>>

Opens given Uniform Resourse Locator

Returns Task()

pub(crate) fn on_register(&mut self) -> Task<Action<Message>>

Sets state as busy and sets which finger is being registered for subscription

Returns Task()

pub(crate) fn on_update_config( &mut self, config: Config, ) -> Task<Action<Message>>

Sets the config state as the given on and writes it to disk

Returns Task()

impl AppModel

pub(crate) fn view_main(&self) -> Element<'_, Message>

Used to construct the main view of application

Returns column with one or two rows of button widgets

fn finger_button(&self, finger: Finger, height: f32) -> Element<'_, Message>

Constructs custom_image_buttons for the main UI based on given height & Finger

Returns an instance of custom_image_button widget

pub(crate) fn view_old(&self) -> Element<'_, Message>

The first UI version which can still be enabled from Settings

pub(crate) fn view_header(&self) -> Element<'_, Message>

Title for the traditional UI

pub(crate) fn view_finger_picker(&self) -> Option<Element<'_, Message>>

Generates a dropdown menu from which to choose which finger is registered

Returns pick_list widget with all Fingers localized names

pub(crate) fn view_icon(&self) -> Element<'_, Message>

Icon for traditional UI

Returns svg widget from FPRINT_ICON

pub(crate) fn view_status(&self) -> Element<'_, Message>

Used to render the current AppModel status in main view

Returns text widget in a container

pub(crate) fn view_progress(&self) -> Option<Element<'_, Message>>

Generates a bar reflecting how many succesful attempts away enrolling print is

Returns progress_bar widget from 0 to num_enroll_steps

pub(crate) fn view_controls(&self) -> Element<'_, Message>

State dependent generation for main controls of the application:

Register, Delete, Verify & Cancel

Returns row widget containing text button widget

impl AppModel

pub fn about(&self) -> Element<'_, Message>

The about page for this app.

pub fn settings(&self) -> Element<'_, Message>

Settings menu

pub(crate) fn list_fingers_task(&self) -> Task<Action<Message>>

Gets all registered prints for requested user

fn update_title_task(&mut self) -> Task<Action<Message>>

Updates the header and window titles.